ePubに対応するよう修正

+++ functions/eyecatch.php
@@ -24,6 +24,9 @@ function pixiv_output($post = null){
     }
 }
 
+
+
+
 /**
  * 投稿がPixivタグを持っているか否かを返す
  *
@@ -35,6 +38,9 @@ function has_pixiv($post = null){
     return (boolean)get_post_meta($post->ID, '_pixiv_embed', true);
 }
 
+
+
+
 /**
  * サムネイル用のメタボックスに表示するHTML
  *
@@ -43,33 +49,63 @@ function has_pixiv($post = null){
  * @return string
  */
 add_filter('admin_post_thumbnail_html', function($content, $post_id){
-	if( 'post' !== get_post_type($post_id) ){
-		return $content;
-	}
-    $help_url = home_url('/faq/pixiv-embed/');
-    $nonce = wp_create_nonce('pixiv_embed_nonce');
-    $embed = esc_textarea(get_post_meta($post_id, '_pixiv_embed', true));
-    $output = <<<HTML
-		<p class="description">または</p>
-		<input type="hidden" name="pixiv_embed_nonce" value="{$nonce}" />
-		<label for="pixiv-embed-tag">Pixivの画像を貼付ける<a href="{$help_url}" target="_blank" data-tooltip-title="Pixivのembedタグについてはこちらをクリック"><i class="dashicons dashicons-editor-help"></i></a></label>
-		<textarea placeholder="埋め込みタグをここにコピペしてください" name="pixiv_embed_tag" id="pixiv_embed_tag">{$embed}</textarea>
+	switch( get_post_type($post_id) ){
+		case 'post':
+		    $help_url = home_url('/faq/pixiv-embed/');
+		    $nonce = wp_create_nonce('pixiv_embed_nonce');
+		    $embed = esc_textarea(get_post_meta($post_id, '_pixiv_embed', true));
+			$content .= <<<HTML
+				<p class="description">または</p>
+				<input type="hidden" name="pixiv_embed_nonce" value="{$nonce}" />
+				<label for="pixiv-embed-tag">Pixivの画像を貼付ける<a href="{$help_url}" target="_blank" data-tooltip-title="Pixivのembedタグについてはこちらをクリック"><i class="dashicons dashicons-editor-help"></i></a></label>
+				<textarea placeholder="埋め込みタグをここにコピペしてください" name="pixiv_embed_tag" id="pixiv_embed_tag">{$embed}</textarea>
 HTML;
-    return $content.$output;
+			break;
+		case 'series':
+			$help_url = home_url('/fat/ebook-cover-regulation');
+			$content .= <<<HTML
+				<p class="description">
+					アイキャッチ画像は電子書籍の表紙として利用されます。サイズは次の通りです（<a href="{$help_url}" target="_blank">もっと詳しく</a>）。
+				</p>
+				<table class="sidemeta__table sidemeta__table--regulation">
+					<tr>
+						<th>横幅</th>
+						<td>1200px</td>
+					</tr>
+					<tr>
+						<th>高さ</th>
+						<td>1920px</td>
+					</tr>
+					<tr>
+						<th>解像度</th>
+						<td>72ppi</td>
+					</tr>
+				</table>
+HTML;
+			break;
+		default:
+			// Do nothing.
+			break;
+	}
+	return $content;
 }, 10, 2);
 
 
 /**
  * PixivのIDを保存する
- * @param type $post_id
+ *
+ * @param int $post_id
+ * @param WP_Post $post
  */
-function _pixiv_embed_save($post_id){
-    if(isset($_REQUEST['pixiv_embed_nonce']) && wp_verify_nonce($_REQUEST['pixiv_embed_nonce'], 'pixiv_embed_nonce')){
+add_action('save_post', function($post_id, WP_Post $post){
+	if( wp_is_post_autosave($post) || wp_is_post_revision($post) ){
+		return;
+	}
+    if( isset($_REQUEST['pixiv_embed_nonce']) && wp_verify_nonce($_REQUEST['pixiv_embed_nonce'], 'pixiv_embed_nonce') ){
         if(isset($_REQUEST['pixiv_embed_tag']) && !empty($_REQUEST['pixiv_embed_tag'])){
             update_post_meta($post_id, '_pixiv_embed', (string)$_REQUEST['pixiv_embed_tag']);
         }else{
             delete_post_meta($post_id, '_pixiv_embed');
         }
     }
-}
-add_action('save_post', '_pixiv_embed_save');
+}, 10, 2);
